rename loop to tracer
