+++ cmake/version.cpp
@@ -1,73 +1,73 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: (c) 2022, Michael Herwig <contact@michael-herwig.de>
-#define LOOP_OBJ
+#define TRACER_OBJ
 
-#include "loop/version.hpp"
+#include "tracer/version.hpp"
 
 // clang-format off
-#define LOOP_VERSION_MAJOR @PROJECT_VERSION_MAJOR@
-#define LOOP_VERSION_MINOR @PROJECT_VERSION_MINOR@
-#define LOOP_VERSION_PATCH @PROJECT_VERSION_PATCH@
-#define LOOP_VERSION_FULL "@PROJECT_VERSION_FULL@"
-#define LOOP_VERSION_CORE "@PROJECT_VERSION@"
-#define LOOP_VERSION_PRE_RELEASE "@PROJECT_VERSION_PRE_RELEASE@"
-#define LOOP_VERSION_BUILD "@PROJECT_VERSION_BUILD@"
-#define LOOP_VERSION_GIT_BRANCH "@GIT_BRANCH@"
-#define LOOP_VERSION_GIT_HASH "@GIT_COMMIT_HASH@"
-#define LOOP_VERSION_GIT_HASH_SHORT "@GIT_COMMIT_HASH_SHORT@"
+#define TRACER_VERSION_MAJOR @PROJECT_VERSION_MAJOR@
+#define TRACER_VERSION_MINOR @PROJECT_VERSION_MINOR@
+#define TRACER_VERSION_PATCH @PROJECT_VERSION_PATCH@
+#define TRACER_VERSION_FULL "@PROJECT_VERSION_FULL@"
+#define TRACER_VERSION_CORE "@PROJECT_VERSION@"
+#define TRACER_VERSION_PRE_RELEASE "@PROJECT_VERSION_PRE_RELEASE@"
+#define TRACER_VERSION_BUILD "@PROJECT_VERSION_BUILD@"
+#define TRACER_VERSION_GIT_BRANCH "@GIT_BRANCH@"
+#define TRACER_VERSION_GIT_HASH "@GIT_COMMIT_HASH@"
+#define TRACER_VERSION_GIT_HASH_SHORT "@GIT_COMMIT_HASH_SHORT@"
 // clang-format on
 
-namespace loop
+namespace tracer
 {
 
 int version_info::major()
 {
-  return LOOP_VERSION_MAJOR;
+  return TRACER_VERSION_MAJOR;
 }
 
 int version_info::minor()
 {
-  return LOOP_VERSION_MINOR;
+  return TRACER_VERSION_MINOR;
 }
 
 int version_info::patch()
 {
-  return LOOP_VERSION_PATCH;
+  return TRACER_VERSION_PATCH;
 }
 
 const char *version_info::full()
 {
-  return LOOP_VERSION_FULL;
+  return TRACER_VERSION_FULL;
 }
 
 const char *version_info::core()
 {
-  return LOOP_VERSION_CORE;
+  return TRACER_VERSION_CORE;
 }
 
 const char *version_info::pre_release()
 {
-  return LOOP_VERSION_PRE_RELEASE;
+  return TRACER_VERSION_PRE_RELEASE;
 }
 
 const char *version_info::build()
 {
-  return LOOP_VERSION_BUILD;
+  return TRACER_VERSION_BUILD;
 }
 
 const char *version_info::git_branch()
 {
-  return LOOP_VERSION_GIT_BRANCH;
+  return TRACER_VERSION_GIT_BRANCH;
 }
 
 const char *version_info::git_hash()
 {
-  return LOOP_VERSION_GIT_HASH;
+  return TRACER_VERSION_GIT_HASH;
 }
 
 const char *version_info::git_hash_short()
 {
-  return LOOP_VERSION_GIT_HASH_SHORT;
+  return TRACER_VERSION_GIT_HASH_SHORT;
 }
 
-} // namespace loop
+} // namespace tracer
