+++ config.hpp
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: (c) 2022, Michael Herwig <contact@michael-herwig.de>
 #pragma once
-#ifndef LOOP_CONFIG_HPP_INCLUDED
-#define LOOP_CONFIG_HPP_INCLUDED
+#ifndef TRACER_CONFIG_HPP_INCLUDED
+#define TRACER_CONFIG_HPP_INCLUDED
 
 // clang-format off
-#cmakedefine LOOP_BUILD_SHARED
-#cmakedefine LOOP_BIG_ENDIAN
+#cmakedefine TRACER_BUILD_SHARED
+#cmakedefine TRACER_BIG_ENDIAN
 // clang-format on
 
-#endif // LOOP_CONFIG_HPP_INCLUDED
+#endif // TRACER_CONFIG_HPP_INCLUDED
