+++ inc/tracer/memory.hpp
@@ -1,14 +1,14 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: (c) 2022, Michael Herwig <contact@michael-herwig.de>
 #pragma once
-#ifndef LOOP_MEMORY_HPP_INCLUDED
-#define LOOP_MEMORY_HPP_INCLUDED
+#ifndef TRACER_MEMORY_HPP_INCLUDED
+#define TRACER_MEMORY_HPP_INCLUDED
 
 #include <cstddef>
 
-#include "loop/api.hpp"
+#include "tracer/api.hpp"
 
-namespace loop
+namespace tracer
 {
 
 #if defined(__cpp_lib_byte) && __cpp_lib_byte >= 201603L
@@ -22,8 +22,8 @@ inline constexpr byte operator""_byte(unsigned long long num)
   return static_cast<byte>(num);
 }
 
-LOOP_API void memzero(byte *data, std::size_t datalen);
+TRACER_API void memzero(byte *data, std::size_t datalen);
 
-} // namespace loop
+} // namespace tracer
 
-#endif // LOOP_MEMORY_HPP_INCLUDED
+#endif // TRACER_MEMORY_HPP_INCLUDED
